Move history code to dedicated package
