+++ history/history_component.go
@@ -1,4 +1,4 @@
-package emqutiti
+package history
 
 import (
 	"strings"
