+++ history/history_select_test.go
@@ -1,4 +1,4 @@
-package emqutiti
+package history
 
 import (
 	"testing"
