+++ history/history_util.go
@@ -1,4 +1,4 @@
-package emqutiti
+package history
 
 import "github.com/charmbracelet/bubbles/list"
