+++ history/historystore.go
@@ -1,4 +1,4 @@
-package emqutiti
+package history
 
 import (
 	"encoding/json"
