+++ history/historystore_query_test.go
@@ -1,4 +1,4 @@
-package emqutiti
+package history
 
 import (
 	"reflect"
