+++ history/historystore_search_test.go
@@ -1,4 +1,4 @@
-package emqutiti
+package history
 
 import (
 	"testing"
